+++ src/mocks/handlers/index.ts
@@ -1,3 +1,3 @@
-import { getJoinedSpacesHandler } from '@/mocks/handlers/spaces/getSpacesHandler';
+import { getJoinedSpacesHandler } from '@/mocks/handlers/spaces/get-joined-spaces-handler';
 
 export const handlers = [getJoinedSpacesHandler];
