+++ src/mocks/handlers/spaces/get-joined-spaces-handler.ts
